requirements.txt
aiofiles==0.5.0
discord.py==1.3.3
jinja2==2.11.2
loguru==0.5.1